Added playlist summary fields

+++ src/application/model/MediaPlaylist.java
@@ -42,7 +42,7 @@ public class MediaPlaylist {
     public int getCompletedCount() {return completedCount;}
     public int getInProgressCount() {return inProgressCount;}
     public int getPlannedCount() {return plannedCount;}
-    public int getAvgRating() {return avgRating;}
+    public double getAvgRating() {return avgRating;}
 	
 	public void setTitle(String title) {this.title = title;}
 	public void setMedias(List<Media> medias) {this.medias = medias;}
